app_V3/requirements_V3.txt
faster-whisper==1.1.1
ffmpeg-python==0.2.0
PyYAML==6.0.2
tqdm==4.67.1
numpy==2.2.4
